rename get-rule-tester
