+++ packages/eslint-plugin-mobx/__tests__/no-anonymous-observer.js
@@ -1,4 +1,4 @@
-import { getRuleTester } from "./utils/getRuleTester";
+import { getRuleTester } from "./utils/get-rule-tester";
 
 import rule from "../src/no-anonymous-observer.js"
