+++ packages/eslint-plugin-mobx/__tests__/utils/get-rule-tester.js
